backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pydantic-settings==2.7.1
email-validator==2.2.0
PyJWT==2.10.1
python-multipart==0.0.20
pg8000==1.31.2
celery==5.4.0
pytest==8.3.4
httpx==0.28.1
